Using onMouseDown instead of onClick for editing
Where modals / overlays / editing popups are concerned, using onMouseDown to close them. The main reason for this is due because on the AudioAnalyzer, the user may be using the mouse to change a value, release their click on the outside of the panel and it will close the panel. This stops that from happening